Fixed issue where cost of route increased in non-dynamic case.
A router could report a higher cost to another in a case where the
former was reached via the same port as a third router that was actually
doing the sending to the second.

For example, R4 -> R5 -> R2 -> R1, where a link also exists between R4
and R1. R1 reports a higher cost to reach R2 and this cost is updated
since R1 is reached through the port used to reach R5.