Added support for ``create'' events.
In order to maintain time-ordering of the buffers for each port, a
packet for a flow that starts in the future cannot be added when the
router is initialized. A create event handles this case by taking a
specified packet and adding it to the outgoing queue at the proper time,
and then follows with a send event.

Changed how Router instances send packets with routing and cost
information. A routing algorithm can now ``prepare'' a packet by adding
the important data.