Identified issue causing discrepancy in output.
When multiple events exist for the same device at the same time, the
ordering of these events on the buffer is known, but unknown on the
event queue. Therefore, by choosing to set the packet of the event as
the packet popped off from the head of the queue ensures that they
agree.