Records unacknowledged packets in Flow class.
Maintains list of outstanding packet sequence numbers. When another
packet is sent, the sequence number is added to the end of the list.
When a packet is acknowledged, then the sequence number is removed from
the list.

Note that because packet k will always be sent before k + 1, this list
is automatically in ascending order.

In this manner, three duplicate acknowledgments can be detected as a
difference between the first and second elements that is greater than
three.